+++ game/ThrowingAxe.cpp
@@ -24,7 +24,7 @@ ThrowingAxe::~ThrowingAxe()
 
 void ThrowingAxe::Update(DWORD dt, vector<LPGAMEOBJECT>* coObjects)
 {
-	if (!camera->checkObjectInCamera(x,y,GetWidth(),GetHeight())) // xử lí ra khỏi cam thì kết thúc
+	if (!camera->checkObjectInCamera(x,y, (float)GetWidth(), (float)GetHeight())) // xử lí ra khỏi cam thì kết thúc
 	{ 
 		isFinish = true;
